Update tests and split out into separate folder
