+++ eventkit_cloud/ui/static/ui/app/components/CreateDataPack/ExportAOI.js
@@ -5,10 +5,10 @@ import ol from 'openlayers';
 import css from '../../styles/ol3map.css';
 import {Toolbar, ToolbarGroup, ToolbarSeparator,ToolbarTitle} from 'material-ui/Toolbar';
 import AoiInfobar from './AoiInfobar.js';
-import SearchAOIToolbar from './SearchAOIToolbar.js';
-import DrawAOIToolbar from './DrawAOIToolbar.js';
-import InvalidDrawWarning from './InvalidDrawWarning.js';
-import DropZone from './DropZone.js';
+import SearchAOIToolbar from '../MapTools/SearchAOIToolbar.js';
+import DrawAOIToolbar from '../MapTools/DrawAOIToolbar.js';
+import InvalidDrawWarning from '../MapTools/InvalidDrawWarning.js';
+import DropZone from '../MapTools/DropZone.js';
 import {updateMode, updateAoiInfo, clearAoiInfo, stepperNextDisabled, stepperNextEnabled} from '../../actions/exportsActions.js';
 import {getGeocode} from '../../actions/searchToolbarActions';
 import {processGeoJSONFile, resetGeoJSONFile} from '../../actions/mapToolActions';
