+++ eventkit_cloud/ui/static/ui/app/components/MapTools/DropZone.js
