+++ eventkit_cloud/ui/static/ui/app/components/MapTools/DropZoneDialog.js
