+++ eventkit_cloud/ui/static/ui/app/tests/MapTools/DrawAOIToolbar.spec.js
@@ -1,11 +1,11 @@
-import {DrawAOIToolbar} from '../../components/CreateDataPack/DrawAOIToolbar';
+import {DrawAOIToolbar} from '../../components/MapTools/DrawAOIToolbar';
 import React from 'react';
 import sinon from 'sinon';
 import {mount, shallow} from 'enzyme';
-import {DrawBoxButton} from '../../components/CreateDataPack/DrawBoxButton';
-import {DrawFreeButton} from '../../components/CreateDataPack/DrawFreeButton';
-import {MapViewButton} from '../../components/CreateDataPack/MapViewButton';
-import {ImportButton} from '../../components/CreateDataPack/ImportButton';
+import {DrawBoxButton} from '../../components/MapTools/DrawBoxButton';
+import {DrawFreeButton} from '../../components/MapTools/DrawFreeButton';
+import {MapViewButton} from '../../components/MapTools/MapViewButton';
+import {ImportButton} from '../../components/MapTools/ImportButton';
 import {fakeStore} from '../../__mocks__/fakeStore';
 import { Provider } from 'react-redux';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
@@ -16,18 +16,25 @@ describe('DrawAOIToolbar component', () => {
 
     it('should render a toolbar title and 4 sub components', () => {
         const props = {
+            toolbarIcons: {},
+            updateMode: () => {},
             handleCancel: (sender) => {},
             setMapView: () => {},
             setAllButtonsDefault: sinon.spy(),
+            setBoxButtonSelected: () => {},
+            setFreeButtonSelected: () => {},
+            setMapViewButtonSelected: () => {},
+            setImportButtonSelected: () => {},
+            setImportModalState: () => {},
         }
         const store = fakeStore({});
         const wrapper = mount(<Provider store={store}><DrawAOIToolbar {...props}/></Provider>, {
             context: {muiTheme},
             childContextTypes: {muiTheme: React.PropTypes.object}
         });
-        expect(wrapper.find('.drawButtonsContainer')).toHaveLength(1);
-        expect(wrapper.find('.drawButtonsTitle')).toHaveLength(1);
-        expect(wrapper.find('.drawButtonsTitle').text()).toEqual('TOOLS');
+        expect(wrapper.find('#container')).toHaveLength(1);
+        expect(wrapper.find('#title')).toHaveLength(1);
+        expect(wrapper.find('#title').text()).toEqual('TOOLS');
         expect(wrapper.find(DrawBoxButton)).toHaveLength(1);
         expect(wrapper.find(DrawFreeButton)).toHaveLength(1);
         expect(wrapper.find(MapViewButton)).toHaveLength(1);
