+++ eventkit_cloud/ui/static/ui/app/tests/MapTools/InvalidDrawWarning.spec.js
@@ -1,4 +1,4 @@
-import {InvalidDrawWarning} from '../../components/CreateDataPack/InvalidDrawWarning';
+import {InvalidDrawWarning} from '../../components/MapTools/InvalidDrawWarning';
 import React from 'react';
 import sinon from 'sinon';
 import {mount} from 'enzyme';
